server_django/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
python-decouple==3.6
sqlparse==0.4.3
Pillow==9.3.0
psycopg2==2.9.5
